Upload files for my simple data analysis project.
The file here is our case study in our subject General Ecology Lecture. Here, as the data analyst in our team, I highlighted the production volume of tomatoes in the Philippines per year and the post-harvest losses in a given year just to give idea on how much tomatoes are produced and wasted at the same time. The calculations that you will read later is my work. Our data here comes from the other study of SEARCA researchers, and we adopted it to support our case study. Thank you!